requirements.txt
BeautifulSoup4
mammoth
GitPython
backoff